Moved TargetPropertyName tests to its own package. Moved test domain model classes out of the hiding TargetPropertyNameModel interface. Renamed TargetPropertyNameModel to DomainModel since it is still used in tests.
