+++ processor/src/test/java/org/mapstruct/ap/test/conditional/basic/ConditionalMappingTest.java
@@ -25,8 +25,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @IssueKey("2051")
 @WithClasses({
     BasicEmployee.class,
-    BasicEmployeeDto.class,
-    TargetPropertyNameModel.class
+    BasicEmployeeDto.class
 })
 public class ConditionalMappingTest {
 
@@ -118,253 +117,6 @@ public class ConditionalMappingTest {
 
         employee = mapper.map( new BasicEmployeeDto( "    " ), utils );
         assertThat( employee.getName() ).isNull();
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodInMapperWithTargetPropertyName.class
-    })
-    public void conditionalMethodInMapperWithTargetPropertyName() {
-        ConditionalMethodInMapperWithTargetPropertyName mapper
-            = ConditionalMethodInMapperWithTargetPropertyName.INSTANCE;
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setFirstName( "  " );
-        employeeDto.setLastName( "Testirovich" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = mapper.map( employeeDto );
-        assertThat( employee.getLastName() ).isNull();
-        assertThat( employee.getFirstName() ).isNull();
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodForCollectionMapperWithTargetPropertyName.class
-    })
-    public void conditionalMethodForCollectionMapperWithTargetPropertyName() {
-        ConditionalMethodForCollectionMapperWithTargetPropertyName mapper
-            = ConditionalMethodForCollectionMapperWithTargetPropertyName.INSTANCE;
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setFirstName( "  " );
-        employeeDto.setLastName( "Testirovich" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = mapper.map( employeeDto );
-        assertThat( employee.getLastName() ).isNull();
-        assertThat( employee.getFirstName() ).isNull();
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() ).isNull();
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodInUsesMapperWithTargetPropertyName.class
-    })
-    public void conditionalMethodInUsesMapperWithTargetPropertyName() {
-        ConditionalMethodInUsesMapperWithTargetPropertyName mapper
-            = ConditionalMethodInUsesMapperWithTargetPropertyName.INSTANCE;
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setFirstName( "  " );
-        employeeDto.setLastName( "Testirovich" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = mapper.map( employeeDto );
-        assertThat( employee.getLastName() ).isNull();
-        assertThat( employee.getFirstName() ).isNull();
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodInMapperWithAllOptions.class
-    })
-    public void conditionalMethodInMapperWithAllOptions() {
-        ConditionalMethodInMapperWithAllOptions mapper
-            = ConditionalMethodInMapperWithAllOptions.INSTANCE;
-
-        ConditionalMethodInMapperWithAllOptions.PresenceUtils utils =
-            new ConditionalMethodInMapperWithAllOptions.PresenceUtils();
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setFirstName( "  " );
-        employeeDto.setLastName( "Testirovich" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = new TargetPropertyNameModel.Employee();
-        mapper.map( employeeDto, employee, utils );
-        assertThat( employee.getLastName() ).isNull();
-        assertThat( employee.getFirstName() ).isNull();
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-        assertThat( utils.visited )
-            .containsExactlyInAnyOrder( "firstName", "lastName", "title", "country" );
-        assertThat( utils.visitedSources ).containsExactly( "EmployeeDto" );
-        assertThat( utils.visitedTargets ).containsExactly( "Employee" );
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodInMapperWithAllExceptTarget.class
-    })
-    public void conditionalMethodInMapperWithAllExceptTarget() {
-        ConditionalMethodInMapperWithAllExceptTarget mapper
-            = ConditionalMethodInMapperWithAllExceptTarget.INSTANCE;
-
-        ConditionalMethodInMapperWithAllExceptTarget.PresenceUtils utils =
-            new ConditionalMethodInMapperWithAllExceptTarget.PresenceUtils();
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setFirstName( "  " );
-        employeeDto.setLastName( "Testirovich" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = mapper.map( employeeDto, utils );
-        assertThat( employee.getLastName() ).isEqualTo( "Testirovich" );
-        assertThat( employee.getFirstName() ).isEqualTo( "  " );
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-        assertThat( utils.visited )
-            .containsExactlyInAnyOrder( "firstName", "lastName", "title", "country", "street" );
-        assertThat( utils.visitedSources ).containsExactlyInAnyOrder( "EmployeeDto", "AddressDto" );
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ConditionalMethodWithTargetPropertyNameInContextMapper.class
-    })
-    public void conditionalMethodWithTargetPropertyNameInUsesContextMapper() {
-        ConditionalMethodWithTargetPropertyNameInContextMapper mapper
-            = ConditionalMethodWithTargetPropertyNameInContextMapper.INSTANCE;
-
-        ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtils utils =
-            new ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtils();
-
-        TargetPropertyNameModel.EmployeeDto employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setLastName( "  " );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        TargetPropertyNameModel.Employee employee = mapper.map( employeeDto, utils );
-        assertThat( employee.getLastName() ).isNull();
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-        assertThat( utils.visited )
-            .containsExactlyInAnyOrder( "firstName", "lastName", "title", "country", "street" );
-
-        ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtilsAllProps allPropsUtils =
-            new ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtilsAllProps();
-
-        employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setLastName( "Tester" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        employee = mapper.map( employeeDto, allPropsUtils );
-        assertThat( employee.getLastName() ).isEqualTo( "Tester" );
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 6" );
-        assertThat( allPropsUtils.visited )
-            .containsExactlyInAnyOrder(
-                "firstName",
-                "lastName",
-                "title",
-                "country",
-                "active",
-                "age",
-                "boss",
-                "primaryAddress",
-                "addresses",
-                "street"
-            );
-
-        ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtilsAllPropsWithSource allPropsUtilsWithSource =
-            new ConditionalMethodWithTargetPropertyNameInContextMapper.PresenceUtilsAllPropsWithSource();
-
-        TargetPropertyNameModel.EmployeeDto bossEmployeeDto = new TargetPropertyNameModel.EmployeeDto();
-        bossEmployeeDto.setLastName( "Boss Tester" );
-        bossEmployeeDto.setCountry( "US" );
-        bossEmployeeDto.setAddresses( Collections.singletonList( new TargetPropertyNameModel.AddressDto(
-            "Testing St. 10" ) ) );
-
-        employeeDto = new TargetPropertyNameModel.EmployeeDto();
-        employeeDto.setLastName( "Tester" );
-        employeeDto.setCountry( "US" );
-        employeeDto.setBoss( bossEmployeeDto );
-        employeeDto.setAddresses(
-            Collections.singletonList( new TargetPropertyNameModel.AddressDto( "Testing St. 6" ) )
-        );
-
-        employee = mapper.map( employeeDto, allPropsUtilsWithSource );
-        assertThat( employee.getLastName() ).isEqualTo( "Tester" );
-        assertThat( employee.getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getAddresses() ).isNotEmpty();
-        assertThat( employee.getAddresses().get( 0 ).getStreet() ).isEqualTo( "Testing St. 6" );
-        assertThat( employee.getBoss() ).isNotNull();
-        assertThat( employee.getBoss().getCountry() ).isEqualTo( "US" );
-        assertThat( employee.getBoss().getLastName() ).isEqualTo( "Boss Tester" );
-        assertThat( employee.getBoss().getAddresses() )
-            .extracting( TargetPropertyNameModel.Address::getStreet )
-            .containsExactly( "Testing St. 10" );
-        assertThat( allPropsUtilsWithSource.visited )
-            .containsExactly(
-                "firstName",
-                "lastName",
-                "title",
-                "country",
-                "active",
-                "age",
-                "boss",
-                "boss.firstName",
-                "boss.lastName",
-                "boss.title",
-                "boss.country",
-                "boss.active",
-                "boss.age",
-                "boss.boss",
-                "boss.primaryAddress",
-                "boss.addresses",
-                "boss.addresses.street",
-                "primaryAddress",
-                "addresses",
-                "addresses.street"
-            );
     }
 
     @ProcessorTest
