+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/ConditionalMethodForCollectionMapperWithTargetPropertyName.java
@@ -3,26 +3,33 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conditional.basic;
+package org.mapstruct.ap.test.conditional.targetpropertyname;
 
 import org.mapstruct.Condition;
 import org.mapstruct.Mapper;
 import org.mapstruct.TargetPropertyName;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.Employee;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.EmployeeDto;
 import org.mapstruct.factory.Mappers;
 
+import java.util.Collection;
+
 /**
- * @author Filip Hrisafov
  * @author Nikola Ivačič
  */
 @Mapper
-public interface ConditionalMethodInMapperWithTargetPropertyName {
+public interface ConditionalMethodForCollectionMapperWithTargetPropertyName {
 
-    ConditionalMethodInMapperWithTargetPropertyName INSTANCE
-            = Mappers.getMapper( ConditionalMethodInMapperWithTargetPropertyName.class );
+    ConditionalMethodForCollectionMapperWithTargetPropertyName INSTANCE
+            = Mappers.getMapper( ConditionalMethodForCollectionMapperWithTargetPropertyName.class );
 
     Employee map(EmployeeDto employee);
+
+    @Condition
+    default <T> boolean isNotEmpty(Collection<T> collection, @TargetPropertyName String propName) {
+        if ( "addresses".equalsIgnoreCase( propName ) ) {
+            return false;
+        }
+        return collection != null && !collection.isEmpty();
+    }
 
     @Condition
     default boolean isNotBlank(String value, @TargetPropertyName String propName) {
