+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/ConditionalMethodInMapperWithAllExceptTarget.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conditional.basic;
+package org.mapstruct.ap.test.conditional.targetpropertyname;
 
 import org.mapstruct.Condition;
 import org.mapstruct.Context;
@@ -29,11 +29,11 @@ public interface ConditionalMethodInMapperWithAllExceptTarget {
         Set<String> visitedSources = new LinkedHashSet<>();
     }
 
-    TargetPropertyNameModel.Employee map(TargetPropertyNameModel.EmployeeDto employee, @Context PresenceUtils utils);
+    Employee map(EmployeeDto employee, @Context PresenceUtils utils);
 
     @Condition
     default boolean isNotBlank(String value,
-                               TargetPropertyNameModel source,
+                               DomainModel source,
                                @TargetPropertyName String propName,
                                @Context PresenceUtils utils) {
         utils.visited.add( propName );
