+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/ConditionalMethodInMapperWithAllOptions.java
@@ -3,15 +3,13 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conditional.basic;
+package org.mapstruct.ap.test.conditional.targetpropertyname;
 
 import org.mapstruct.Condition;
 import org.mapstruct.Context;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.TargetPropertyName;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.Employee;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.EmployeeDto;
 import org.mapstruct.factory.Mappers;
 
 import java.util.LinkedHashSet;
@@ -39,8 +37,8 @@ public interface ConditionalMethodInMapperWithAllOptions {
 
     @Condition
     default boolean isNotBlank(String value,
-                               TargetPropertyNameModel source,
-                               @MappingTarget TargetPropertyNameModel target,
+                               DomainModel source,
+                               @MappingTarget DomainModel target,
                                @TargetPropertyName String propName,
                                @Context PresenceUtils utils) {
         utils.visited.add( propName );
