+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/ConditionalMethodInUsesMapperWithTargetPropertyName.java
@@ -3,13 +3,11 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conditional.basic;
+package org.mapstruct.ap.test.conditional.targetpropertyname;
 
 import org.mapstruct.Condition;
 import org.mapstruct.Mapper;
 import org.mapstruct.TargetPropertyName;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.Employee;
-import org.mapstruct.ap.test.conditional.basic.TargetPropertyNameModel.EmployeeDto;
 import org.mapstruct.factory.Mappers;
 
 /**
