+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/DomainModel.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright MapStruct Authors.
+ *
+ * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package org.mapstruct.ap.test.conditional.targetpropertyname;
+
+/**
+ * Target Property Name test entities
+ *
+ * @author Nikola Ivačič
+ */
+public interface DomainModel {
+
+}
