+++ processor/src/test/java/org/mapstruct/ap/test/conditional/targetpropertyname/Employee.java
@@ -1,0 +1,99 @@
+/*
+ * Copyright MapStruct Authors.
+ *
+ * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package org.mapstruct.ap.test.conditional.targetpropertyname;
+
+import java.util.List;
+
+/**
+ * @author Nikola Ivačič
+ */
+public class Employee implements DomainModel {
+
+  private String firstName;
+  private String lastName;
+  private String title;
+  private String country;
+  private boolean active;
+  private int age;
+
+  private org.mapstruct.ap.test.conditional.targetpropertyname.Employee boss;
+
+  private Address primaryAddress;
+
+  private List<Address> addresses;
+
+  public String getFirstName() {
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public String getLastName() {
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+
+  public String getTitle() {
+    return title;
+  }
+
+  public void setTitle(String title) {
+    this.title = title;
+  }
+
+  public String getCountry() {
+    return country;
+  }
+
+  public void setCountry(String country) {
+    this.country = country;
+  }
+
+  public boolean isActive() {
+    return active;
+  }
+
+  public void setActive(boolean active) {
+    this.active = active;
+  }
+
+  public int getAge() {
+    return age;
+  }
+
+  public void setAge(int age) {
+    this.age = age;
+  }
+
+  public org.mapstruct.ap.test.conditional.targetpropertyname.Employee getBoss() {
+    return boss;
+  }
+
+  public void setBoss(org.mapstruct.ap.test.conditional.targetpropertyname.Employee boss) {
+    this.boss = boss;
+  }
+
+  public Address getPrimaryAddress() {
+    return primaryAddress;
+  }
+
+  public void setPrimaryAddress(Address primaryAddress) {
+    this.primaryAddress = primaryAddress;
+  }
+
+  public List<Address> getAddresses() {
+    return addresses;
+  }
+
+  public void setAddresses(List<Address> addresses) {
+    this.addresses = addresses;
+  }
+}
